Fix compilation on 32-bit systems like armv7l
On some 32-bits targets with glibc, time_t still defaults to 32-bit (on
armv7l: a `long`). Meanwhile, a `SQLiteWriter::var_t` can hold either an
`int`, a `long long`, or a `long long unsigned`. Resolve this by casting
to `int64_t`, which is a `long long`.

If compiled for targets where time_t is already a 64-bit integer, this
has no effect.

+++ notifiers.cc
@@ -163,7 +163,7 @@ void Notifier::bulkAlert(const std::string& textBody)
 void SQLiteWriterNotifier::alert(const std::string& str)
 {
   if(g_sqlw)
-    g_sqlw->addValue({{"tstamp", time(nullptr)}, {"message", str}}, "notifications");
+    g_sqlw->addValue({{"tstamp", (int64_t)time(nullptr)}, {"message", str}}, "notifications");
 }
 
 void Notifier::bulkDone()
